微信公共号

+++ starcloud-llmops/starcloud-business/starcloud-business-open/starcloud-business-open-biz/src/main/java/com/starcloud/ops/business/open/service/impl/WechatServiceImpl.java
@@ -235,6 +235,9 @@ public class WechatServiceImpl implements WechatService {
     @Override
     public void delete(String uid) {
         AppPublishChannelRespVO publishChannelRespVO = appPublishChannelService.get(uid);
+        if (publishChannelRespVO == null) {
+            return;
+        }
         appPublishChannelService.delete(uid);
         BaseChannelConfigDTO config = publishChannelRespVO.getConfig();
         if (config instanceof WeChatAccountChannelConfigDTO) {
